By the time of this commit SignalR finally worked
It was the fact that THE DUMB PIECE OF JUNK WHO MADE THE TUTORIAL I WATCHED DIDN'T OVERRIDE AN IMPORTANT METHOD (the one I overrided in the previous commit).

Why did SignalR work only with on web pages? Because thanks to the notification.js file they used a method in the NotificationUserHub called GetConnectionId(), which contained the actual code that connected them to the server, while the Android client was NOT causing the method to be invoked. So the code from GetConnectionId is in the OnConnectedAsync() as well. That's it!